% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{calibSpectrum}
\alias{calibSpectrum}
\title{calibSpectrum}
\usage{
calibSpectrum(spec, zoneref, ppmref)
}
\arguments{
\item{spec}{a spectrum object returned by the readSpectrum function}

\item{zoneref}{the ppm range containing the TSP/DSS signal}

\item{ppmref}{the ppm reference value}
}
\value{
spec object
}
\description{
\code{calibSpectrum} belongs to the low-level functions group - it processes only one raw spectrum at time.
}
