% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{cleanPeaks}
\alias{cleanPeaks}
\title{cleanPeaks}
\usage{
cleanPeaks(spec, model, SNthreshold = 5)
}
\arguments{
\item{spec}{a 'spec' object}

\item{model}{a model object}

\item{SNthreshold}{Threshold for the Signal-Noise Ratio below which the peaks will be rejected}
}
\value{
a data.frame of the remaining peaks
}
\description{
\code{cleanPeaks} cleans the peaks under a specified threshold
}
