% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkmatchTools.R
\name{clusterMerging}
\alias{clusterMerging}
\title{clusterMerging}
\usage{
clusterMerging(clust1, dfident1, clust2, dfident2)
}
\arguments{
\item{clust1}{object obtained using the \code{getClusters} function}

\item{dfident1}{obtained using the \code{matchClusters} function}

\item{clust2}{object obtained using the \code{getClusters} function}

\item{dfident2}{obtained using the \code{matchClusters} function}
}
\description{
merging 2 clustering approaches
}
