% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{computeBL}
\alias{computeBL}
\title{computeBL}
\usage{
computeBL(spec, model)
}
\arguments{
\item{spec}{a 'spec' object}

\item{model}{a model object}
}
\value{
a vector of the baseline estimated during the deconvolution process
}
\description{
\code{computeBL} computes baseline based on the model.
}
