% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\docType{data}
\name{deconvParams}
\alias{deconvParams}
\title{deconvParams}
\format{
An object of class \code{list} of length 45.
}
\usage{
deconvParams
}
\value{
\itemize{
  \item \code{flist} : Filter type list : 'smooth1', 'smooth2' and'smooth3' for Savitzky-Golay filter, 'daub8' and 'symlet8' for filter based on wavelet
  \item \code{criterion} : Criterion type for the optimizations : 0 => R2, 1 => 1/Std(residues) - default value = 0
  \item \code{reltol} : Criterion tolerance for the optimization - default value = 0.0001
  \item \code{facN} :  Noise factor applied while the peak finding step - default value = NULL
  \item \code{ratioPN} : Peak/Noise Ratio applied while the peak selection step - default value = 1
  \item \code{obl} : Optimization of a baseline (BL) for each massif. 0 means no BL, an integer greater than 0 indicates the polynomial order of the BL default value = 0
  \item \code{distPeaks} : PeakFinder : min distance between 2 peaks (as multiple of sigma_min which is typically equal to 0.0005 ppm) - default value = 2
  \item \code{optim} : Indicates if optimisation is applied - default value = 1
  \item \code{oppm} : Indicates if ppm optimisation is applied - default value = 1
  \item \code{osigma} : Indicates if sigma optimisation is applied - default value = 1
  \item \code{d2meth} : PeakFinder : Indicates if minima method to the second derivation is applied
  \item \code{spcv} : PeakFinder : Maximal CV on Spectrum - default value = 0.005
  \item \code{d2cv} : PeakFinder : Maximum CV on the derived spectrum - default value = 0.05
  \item \code{d1filt} : Apply Filter (1) on the 1st derivate or not (0) - default value = 0
  \item \code{d2filt} : Apply Filter (1) on the 2nd derivate or not (0) - default value = 0
  \item \code{sigma_min} : Optimization of Sigmas : Fixe the minimum limit of sigmas - default value = 0.0005
  \item \code{sigma_max} : Optimization of Sigmas : Fixe the maximum limit of sigmas - default value = 0.005
  \item \code{verbose} : Indicates if we want information messages - default value = 1
  \item \code{exclude_zones} : Exclude ppm zones for the criterion evaluation - default value = NULL
}
}
\description{
Initialize the deconvolution parameter list
}
\keyword{datasets}
