% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doProcessing.R
\name{detectCores}
\alias{detectCores}
\title{detectCores}
\usage{
detectCores(...)
}
\arguments{
\item{...}{See \code{parallel::detectCores}}
}
\description{
\code{detectCores} is simply a shortcut for parallel::detectCores().
}
