% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnmrTools.R
\name{doProcCmd}
\alias{doProcCmd}
\title{doProcCmd}
\usage{
doProcCmd(specObj, cmdstr, ncpu = 1, debug = FALSE)
}
\arguments{
\item{specObj}{a complex list return by \code{doProcessing} function. See the manual page of the \code{\link{doProcessing}} function for more details on its structure.}

\item{cmdstr}{the Macro-commands string array; See the Macro-command Reference Guide (\url{https://nmrprocflow.org/themes/pdf/Macrocommand.pdf}) to have more details about macro-commands.}

\item{ncpu}{The number of cores [default: 1]}

\item{debug}{a boolean to specify if we want the function to be more verbose.}
}
\value{
\code{specMat} : a 'specMat' object - See the manual page of the \code{\link{doProcessing}} function for more details on its structure
}
\description{
\code{doProcCmd} it process the Macro-commands string array specified at input.
}
\examples{
 \dontrun{
    data_dir <- system.file("extra", package = "Rnmr1D")
    RAWDIR <- file.path(data_dir, "MMBBI_14P05")
    CMDFILE <- file.path(data_dir, "NP_macro_cmd.txt")
    SAMPLEFILE <- file.path(data_dir, "Samples.txt")
    out <- Rnmr1D::doProcessing(RAWDIR, cmdfile=CMDFILE, samplefile=SAMPLEFILE, ncpu=detectCores())
# Plot the ppm range [0 .. 3]
    plotSpecMat(ou$specMat, ppm_lim=c(0.5,3), K=0)
# clean some ppm zones 
    specMat.new <- Rnmr1D::doProcCmd(out, 
             c("zero", 
                 "2.71 2.65", 
                 "2.149 2.208", 
                 "1.55 1.48", 
                 "1.265 1.19", 
                 "0.86 0.76", 
                 "EOL"
               "bucket aibin 10.2 10.5 0.3 3 0", 
                 "9.309 4.788", 
                 "4.67 0.615", 
                 "EOL"
              ),ncpu=2, debug=TRUE)
# Plot the ppm range [0 .. 3] after the cleanning
    dev.new();
    plotSpecMat(specMat.new, ppm_lim=c(0.5,3), K=0)
}
}
