% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{estimateBL}
\alias{estimateBL}
\title{estimateBL}
\usage{
estimateBL(spec, ppmrange, WS = 50, NEIGH = 35)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppmrange}{the ppm range in which the baseline will be estimated}

\item{WS}{Size of the window (in number of points) from which a rolling average will be established}

\item{NEIGH}{The minimum window size (in number of points) in which the signal compared to its mean can be considered as belonging to the baseline.}
}
\value{
a vector of the estimated baseline
}
\description{
\code{estimateBL} estimates of the baseline of the spectrum in the corresponding ppm range (based on the C_Estime_LB routine)
}
