% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{filterByWT}
\alias{filterByWT}
\title{filterByWT}
\usage{
filterByWT(s, wavelet, type = 0)
}
\arguments{
\item{s}{the spectral signal as a numerical vector}

\item{wavelet}{the name of the wavelet: haar, daub2, daub4, daub8, symlet2, symlet4, symlet8}

\item{type}{the type of the threshold :  0 for Soft threshold, 1 for Hard threshold}
}
\value{
a vector of the same dimension as the entry one
}
\description{
\code{filterByWT} applies a filtering based on wavelet using the universal threshold
}
