% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{filterSavGol}
\alias{filterSavGol}
\title{filterSavGol}
\usage{
filterSavGol(s, m, nl, nr)
}
\arguments{
\item{s}{the spectral signal as a numerical vector}

\item{m}{the degree of the polynomial filter (integer)}

\item{nl}{width of the sliding window on the left (integer)}

\item{nr}{width of the sliding window on the rigth (integer)}
}
\value{
a vector of the same dimension as the entry one
}
\description{
\code{filterSavGol} applies a Savitzky-Golay filter on a spectral signal.
}
