% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnmrTools.R
\name{getBucketsTable}
\alias{getBucketsTable}
\title{getBucketsTable}
\usage{
getBucketsTable(specObj)
}
\arguments{
\item{specObj}{a complex list return by \code{doProcessing} function. See the manual page of the \code{\link{doProcessing}} function for more details on its structure.}
}
\value{
the buckets table
}
\description{
Generates the buckets table
}
