% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTools.R
\name{getClusters}
\alias{getClusters}
\title{getClusters}
\usage{
getClusters(data, method = "hca", ...)
}
\arguments{
\item{data}{the matrix including the integrations of the areas defined by the buckets (columns) 
on each spectrum (rows)}

\item{method}{Clustering method of the buckets. Either 'corr' for 'correlation' or 'hca' for 
'hierarchical clustering analysis'.}

\item{...}{Depending on the chosen method:
\itemize{
  \item \code{corr} : cval, dC, ncpu
  \item \code{hca} :  vcutusr
}}
}
\value{
\code{getClusters} returns a list containing the following components:
\itemize{
  \item \code{vstats} cxwcsxcx
  \item \code{clusters} fdfdsfsd
  \item \code{clustertab} the associations matrix that gives for each cluster (column 2) the corresponding buckets (column 1)
  \item \code{params} fdfdsfdsf
  \item \code{vcrit} gfdghdfhgfhgf
  \item \code{indxopt} hgfhgfhgfhgfh
}
}
\description{
Bucket Clustering based on either a lower threshold  applied on correlations or a cutting value 
applied on a hierarchical tree of the variables (buckets) generated by an Hierarchical 
Clustering Analysis (HCA)
}
\details{
At the bucketing step (see above),  we have chosen the intelligent bucketing, it means 
that each bucket exact matches with one resonance peak. Thanks to this, the buckets now 
have a strong chemical meaning, since the resonance peaks are the fingerprints of 
chemical compounds. However, to assign a chemical compound, several resonance peaks 
are generally required in 1D 1 H-NMR metabolic profiling. To generate relevant 
clusters (i.e. clusters possibly matching to chemical compounds), two approaches 
have been implemented:
\itemize{
  \item Bucket Clustering based on a lower threshold  applied on correlations
  \item Bucket Clustering based on a hierarchical tree of the variables (buckets) generated 
by an Hierarchical Clustering Analysis (HCA)
}
}
\examples{
 \dontrun{
     outclust <- getClusters(outMat, method='corr', cval=0, dC=0.005, ncpu=2)
 }
}
\references{
{
  Jacob D., Deborde C. and Moing A. (2013) An efficient spectra processing method for metabolite identification from 1H-NMR metabolomics data.
  Analytical and Bioanalytical Chemistry 405(15) 5049-5061 doi: 10.1007/s00216-013-6852-y
 }
}
