% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{getDeconvParams}
\alias{getDeconvParams}
\title{getDeconvParams}
\usage{
getDeconvParams(params = NULL)
}
\arguments{
\item{params}{a list defining some specific parameters for deconvolution}
}
\value{
the resulting list of deconvolution parameters.
}
\description{
\code{getDeconvParams} merges some specific parameters values with the full deconvolution list and return the resulting list. With no parameter as input, it returns the default parameter list.
}
