% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTools.R
\name{getMergedDataset}
\alias{getMergedDataset}
\title{getMergedDataset}
\usage{
getMergedDataset(data, clustObj, onlycluster = FALSE)
}
\arguments{
\item{data}{the matrix including the integrations of the areas defined by the buckets (columns)
on each spectrum (rows)}

\item{clustObj}{a list generated by the \code{getClusters} function}

\item{onlycluster}{boolean - specifies if the merged data matrix at output must only contain 
the merged clusters (TRUE) or if it must also contain the buckets that are not include within a 
cluster (FALSE)}
}
\description{
merged variables for each cluster (based on their average)
}
