% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnmrTools.R
\name{getSnrDataset}
\alias{getSnrDataset}
\title{getSnrDataset}
\usage{
getSnrDataset(specObj, zone_noise = c(10.2, 10.5), ratio = TRUE)
}
\arguments{
\item{specObj}{a complex list return by \code{doProcessing} function. 
See the manual page of the \code{\link{doProcessing}} function for more details on its structure.}

\item{zone_noise}{Specify a ppm range of noisy zone default is c(10.2,10.5)}

\item{ratio}{boolean; TRUE for output Signal-Noise Ratio, or FALSE to output maximum value of 
each bucket and in addition, the estimate noise as a separate column}
}
\value{
the Signal-Noise-Ratio matrix
}
\description{
Generates the Signal-Noise-Ratio dataset
}
\details{
whatever the bucketing approach used, the Signal-to-Noise ratio is a good quality indicator. 
Thus, it is possible to check buckets based on their Signal-to-Noise ratio.
}
