% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnmrTools.R
\name{getSpectraData}
\alias{getSpectraData}
\title{getSnrDataset}
\usage{
getSpectraData(specObj)
}
\arguments{
\item{specObj}{a complex list return by \code{doProcessing} function.}
}
\value{
the spectral data matrix
}
\description{
Generates the spectral data matrix. The first column indicates the value of ppm, then the following columns correspond to spectral data, one column per spectrum.
}
