% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotTools.R
\name{ggplotClusters}
\alias{ggplotClusters}
\title{ggplotClusters}
\usage{
ggplotClusters(data, clustObj)
}
\arguments{
\item{data}{the matrix including the integrations of the areas defined by the buckets (columns)
on each spectrum (rows)}

\item{clustObj}{a list generated by the \code{getClusters} function}
}
\description{
Plots the boxplot of all clusters allowing to have an insight on the clusters distribution.
Plot based on ggplot2
}
