% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotTools.R
\name{ggplotCriterion}
\alias{ggplotCriterion}
\title{ggplotCriterion}
\usage{
ggplotCriterion(clustObj)
}
\arguments{
\item{clustObj}{a list generated by the \code{getClusters} function}
}
\description{
Plots the curves that show the number of clusters, the number of clustered buckets and the 
size of biggest cluster  versus the criterion, namely the correlation threshold for the 'corr' 
method, the cutting value for the 'hca' method.
}
