% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotTools.R
\name{ggplotPlotly}
\alias{ggplotPlotly}
\title{ggplotPlotly}
\usage{
ggplotPlotly(g, width = NULL, height = NULL, textposition = "right")
}
\arguments{
\item{g}{The ggplot2 graph object to be translated into an interactive plotly version}

\item{width}{Width of the plot in pixels (optional, defaults to automatic sizing).}

\item{height}{Height of the plot in pixels (optional, defaults to automatic sizing)}

\item{textposition}{Position of the labels on the graphs relative to the points. Possible values are : 'right', 'left', 'top' or 'buttom'}
}
\description{
Translate 'ggplot2' graphs to an interactive plotly version
}
