% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{GSDeconv}
\alias{GSDeconv}
\title{GSDeconv}
\usage{
GSDeconv(
  spec,
  ppmrange,
  params = NULL,
  filter = "symlet8",
  scset = c(2, 3, 12),
  verbose = 1
)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppmrange}{a ppm range as a list in order to apply the deconvolution}

\item{params}{a list of specific parameters for deconvolution}

\item{filter}{a filter type for filtering the noise and  smoothing the signal}

\item{scset}{a set of scmin values}

\item{verbose}{level of debug information}
}
\value{
a model object
}
\description{
Global Spectra Deconvolution: \code{GSDeconv} belongs to the low-level functions group for deconvolution.
}
