% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{LSDeconv}
\alias{LSDeconv}
\title{LSDeconv}
\usage{
LSDeconv(
  spec,
  ppmrange,
  params = NULL,
  filterset = 1:6,
  oblset = 1:12,
  verbose = 1
)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppmrange}{a ppm range as a list in order to apply the deconvolution}

\item{params}{a list of specific parameters for deconvolution}

\item{filterset}{a set of filter type for filtering the noise and  smoothing the signal}

\item{oblset}{a set of baseline order for fitting}

\item{verbose}{level of debug information}
}
\value{
a model object
}
\description{
Local Spectra Deconvolution: \code{LSDeconv} belongs to the low-level functions group for deconvolution.
}
