% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkmatchTools.R
\name{matchClusters}
\alias{matchClusters}
\title{matchClusters}
\usage{
matchClusters(
  clusters,
  lib = "dbref6",
  score_min = 0.5,
  ppmtol = 0.02,
  method = "all"
)
}
\arguments{
\item{clusters}{object obtained using the \code{getClusters} function}

\item{lib}{either an internal library name (dbref6 or hmdb110), or the path of a DB file}

\item{score_min}{selects only matchs with a score above the value}

\item{ppmtol}{ppm tolerance for seaching peaks}

\item{method}{either 'all' for all query peaks must machted, or 'one' for only one peak must matched}
}
\description{
Match all clusters obtained using the \code{getClusters} function and using the the \code{peakMatching} function.
}
