% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{MultiLSDeconv}
\alias{MultiLSDeconv}
\title{MultiLSDeconv}
\usage{
MultiLSDeconv(
  spec,
  ppmranges = NULL,
  params = NULL,
  filterset = c(7, 9),
  oblset = 0,
  ncpu = 4,
  verbose = 0
)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppmranges}{ppm ranges as a matrix in order to apply the deconvolution, each row specifying a zone}

\item{params}{a list of specific parameters for deconvolution}

\item{filterset}{a set of filter type for filtering the noise and  smoothing the signal (only if the matrix defining peaks not defined in order to find peaks)}

\item{oblset}{a set of baseline order for fitting}

\item{ncpu}{number of CPU for parallel computing}

\item{verbose}{level of debug information}
}
\value{
a model object
}
\description{
Multiple Local Spectra Deconvolution: \code{MultiLSDeconv} belongs to the low-level functions group for deconvolution.
}
