% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{MyFuncTest}
\alias{MyFuncTest}
\title{MyFuncTest}
\usage{
MyFuncTest(spec, ppmrange, filt, peaks, verbose = 1)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppmrange}{a ppm range as a list in order to apply the deconvolution}

\item{filt}{a list of specific parameters for filtering}

\item{peaks}{a list of specific parameters for deconvolution}

\item{verbose}{level of debug information}
}
\value{
a list
}
\description{
\code{MyFuncTest} belongs to the low-level functions group for deconvolution.
}
