% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{MyFuncTest2}
\alias{MyFuncTest2}
\title{MyFuncTest2}
\usage{
MyFuncTest2(spec, n1, n2)
}
\arguments{
\item{spec}{a 'spec' object}

\item{n1}{start index of the spectrum segment to be corrected}

\item{n2}{stop index of the spectrum segment to be corrected}
}
\value{
a vector of the same dimension as spec$int
}
\description{
\code{MyFuncTest2} belongs to the low-level functions group for deconvolution.
}
