% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{optimOneVoigt}
\alias{optimOneVoigt}
\title{optimOneVoigt}
\usage{
optimOneVoigt(X, Y, par)
}
\arguments{
\item{X}{a vector of ppm values}

\item{Y}{a vector of intensities}

\item{par}{a vector of the 3 pseudo-voigt parameters namely: Amplitude, central ppm value, 2 ppm widths at mid-height for mixed lorentizian and gaussian}
}
\value{
a vector of the pseudo-voigt parameters (same size as par)
}
\description{
\code{optimOneVoigt} belongs to the low-level functions group for deconvolution.
}
