% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{peakFiltering}
\alias{peakFiltering}
\title{peakFiltering}
\usage{
peakFiltering(spec, peaks, ratioPN)
}
\arguments{
\item{spec}{a 'spec' object}

\item{peaks}{the matrix defining peaks, one peak by row, with columns defined as : pos, ppm, amp, sigma, eta}

\item{ratioPN}{the ratio Peaks/Noise for filtering}
}
\value{
a dataframe
}
\description{
\code{peakFiltering} belongs to the low-level functions group for deconvolution.
}
