% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{peakFinder}
\alias{peakFinder}
\title{peakFinder}
\usage{
peakFinder(spec, ppmrange, params = NULL, filter = "none", verbose = 1)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppmrange}{a ppm range as a list in order to apply the deconvolution}

\item{params}{a list of specific parameters for deconvolution}

\item{filter}{a filter type for filtering the noise and  smoothing the signal}

\item{verbose}{level of debug information}
}
\value{
a list
}
\description{
\code{peakFinder} belongs to the low-level functions group for deconvolution.
}
