% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkmatchTools.R
\name{peakMatching}
\alias{peakMatching}
\title{PeakMatching}
\usage{
peakMatching(peaklist, lib = "dbref6", ppmtol = 0.02, method = "all")
}
\arguments{
\item{peaklist}{a list of ppm, each corresponding to the maximal intensity of a compound peak}

\item{lib}{either an internal library name (dbref6 or hmdb110), or the path of a DB file}

\item{ppmtol}{ppm tolerance for seaching peaks}

\item{method}{either 'all' for all query peaks must machted, or 'one' for only one peak must matched}
}
\description{
Searching for putative compounds in a peaklist database from a list of chemical shifts,
based for example on a clustering obtained using the \code{getClusters} function.
}
