% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{peakOptimize}
\alias{peakOptimize}
\title{peakOptimize}
\usage{
peakOptimize(spec, ppmrange, peaks, verbose = 1)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppmrange}{a ppm range as a list in order to apply the deconvolution}

\item{peaks}{a list of specific parameters for deconvolution, including the matrix defining peaks, one peak by row, with columns defined as : pos, ppm, amp, sigma, pfac, integral}

\item{verbose}{level of debug information}
}
\value{
a list
}
\description{
\code{peakOptimize} belongs to the low-level functions group for deconvolution.
}
