% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTools.R
\name{plotClusters}
\alias{plotClusters}
\title{plotClusters}
\usage{
plotClusters(data, clustObj, horiz = TRUE,
  main = "Boxplot by clusters (log10 transformed)")
}
\arguments{
\item{data}{the matrix including the integrations of the areas defined by the buckets (columns)
on each spectrum (rows)}

\item{clustObj}{a list generated by the \code{getClusters} function}

\item{horiz}{Boolean - Indicates if the plot is horizontal (TRUE) or vertical (FALSE)}

\item{main}{Main title of the plot}
}
\description{
Plots the boxplot of all clusters allowing to have an insight on the clusters distribution
}
