% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{plotModel}
\alias{plotModel}
\title{plotModel}
\usage{
plotModel(spec, model, exclude_zones = NULL, title = "")
}
\arguments{
\item{spec}{a 'spec' object (see \code{readSpectrum}, \code{Spec1rDoProc})}

\item{model}{a 'model' object (see \code{specDeconv}, \code{peakOptimize}, \code{GSDeconv}, \code{LSDeconv})}

\item{exclude_zones}{a list of vector defining the excluded zones for lorentzian plots}

\item{title}{title of the graphic}
}
\description{
\code{plotModel} plots the model along with the resulting lorentzians from deconvolution
}
