% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{plotModelwithResidus}
\alias{plotModelwithResidus}
\title{plotModelwithResidus}
\usage{
plotModelwithResidus(
  spec,
  model,
  ynames = c("Origin", "Model", "Residues"),
  title = ""
)
}
\arguments{
\item{spec}{a 'spec' object (see \code{readSpectrum}, \code{Spec1rDoProc})}

\item{model}{a 'model' object (see \code{specDeconv}, \code{peakOptimize}, \code{GSDeconv}, \code{LSDeconv})}

\item{ynames}{a vector defining the y names. default value = c('Origin', 'Model', 'Residues')}

\item{title}{title of the graphic}
}
\description{
\code{plotSpec} plots the model along with the resulting residues from deconvolution
}
