% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{plotSpec}
\alias{plotSpec}
\title{plotSpec}
\usage{
plotSpec(
  spec,
  ppm,
  x,
  y,
  ynames = c("Origin", "Filtered", "Model"),
  ycolors = c("grey", "blue", "red", "green", "orange", "magenta", "cyan", "darkgreen",
    "darkorange"),
  title = ""
)
}
\arguments{
\item{spec}{a 'spec' object (see \code{readSpectrum}, \code{Spec1rDoProc})}

\item{ppm}{a ppm range defining the window of the plotting}

\item{x}{a vector defining the x-axis (abscissa)}

\item{y}{a vector or a matrix defining the y-axes (ordinates), each signal as a column}

\item{ynames}{a vector defining the y names (same order as the y matrix)}

\item{ycolors}{a vector defining the y colors (same order as the y matrix)}

\item{title}{title of the graphic}
}
\description{
\code{plotSpec} plots all signals defined by a matrix.
}
