% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{pos2ppm}
\alias{pos2ppm}
\title{pos2ppm}
\usage{
pos2ppm(spec, index)
}
\arguments{
\item{spec}{a 'spec' object}

\item{index}{an index position}
}
\value{
the corresponding ppm value
}
\description{
\code{pos2ppm} convert an index position to the corresponding ppm value
}
