% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{ppm2pos}
\alias{ppm2pos}
\title{ppm2pos}
\usage{
ppm2pos(spec, ppm)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppm}{a ppm value}
}
\value{
the corresponding index position
}
\description{
\code{ppm2pos} convert a ppm value to the corresponding index position
}
