% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{data}
\name{PPM_MAX}
\alias{PPM_MAX}
\title{PPM_MAX}
\format{An object of class \code{numeric} of length 1.}
\usage{
PPM_MAX
}
\description{
Maximal ppm value for 1H NMR to consider in the processing step and then to store in the specMat object
Default value = 11
}
\keyword{datasets}
