% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{data}
\name{PPM_MAX_13C}
\alias{PPM_MAX_13C}
\title{PPM_MAX_13C}
\format{An object of class \code{numeric} of length 1.}
\usage{
PPM_MAX_13C
}
\description{
Maximal ppm value for 13C NMR to consider in the processing step and then to store in the specMat object
Default value = 200
}
\keyword{datasets}
