% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{data}
\name{PPM_MIN}
\alias{PPM_MIN}
\title{PPM_MIN}
\format{An object of class \code{numeric} of length 1.}
\usage{
PPM_MIN
}
\description{
Minimal ppm value for 1H NMR to consider in the processing step and then to store in the specMat object
Default value = 0.5
}
\keyword{datasets}
