% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{PVoigt}
\alias{PVoigt}
\title{PVoigt}
\usage{
PVoigt(ppm, amp, x0, sigma, eta)
}
\arguments{
\item{ppm}{a vector of ppm values}

\item{amp}{amplitude of the lorentzian}

\item{x0}{central value of the lorentzian}

\item{sigma}{half-width of both lorentzian and gaussian}

\item{eta}{mixing coefficient for the pseudo-voigt function (between 0 and 1)}
}
\value{
a vector of the lorentzian values (same size as ppm)
}
\description{
\code{PVoigt} belongs to the low-level functions group for deconvolution.
}
