% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{readSpectrum}
\alias{readSpectrum}
\title{readSpectrum}
\usage{
readSpectrum(
  ACQDIR,
  procParams,
  ppmnoise = c(10.2, 10.5),
  PHC = NULL,
  scaleIntensity = 1,
  verbose = 1
)
}
\arguments{
\item{ACQDIR}{Full directory path of the raw spectrum, i.e the directory containing the FID}

\item{procParams}{a Spec1rProcpar list}

\item{ppmnoise}{the ppm range containing only noise signal in order to estimate the level of the noise (S/N ratio)}

\item{PHC}{the phasing values applied to the spectrum in the frequency domain thus avoiding the automatic phasing step. Only useful if the input signal is an FID (procParams$INPUT_SIGNAL)}

\item{scaleIntensity}{factor applied to the intensities to establish a change of scale.}

\item{verbose}{level of debug information}
}
\value{
spec object
}
\description{
\code{read_spectrum} belongs to the low-level functions group - it processes only one raw spectrum at time.
}
