% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnmrTools.R
\name{RWrapperCMD1D}
\alias{RWrapperCMD1D}
\title{RWrapperCMD1D}
\usage{
RWrapperCMD1D(cmdName, specMat, ...)
}
\arguments{
\item{cmdName}{the name of internal function}

\item{specMat}{a 'specMat' object}

\item{...}{specific parameters of the requested function}
}
\value{
\code{specMat} : a 'specMat' object
}
\description{
\code{RWrapperCMD1D} belongs to the low-level functions group - it serves as a wrapper to 
call internale functions for processing.
}
