% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{data}
\name{sep}
\alias{sep}
\title{sep}
\format{An object of class \code{character} of length 1.}
\usage{
sep
}
\description{
Separator value - default value = Tab separator
}
\keyword{datasets}
