% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doProcessing.R
\name{setPPMbounds}
\alias{setPPMbounds}
\title{setPPMbounds}
\usage{
setPPMbounds(proton = c(-0.5, 11), carbon = c(0, 200))
}
\arguments{
\item{proton}{Minimal and Maximal ppm value for 1H NMR}

\item{carbon}{Minimal and Maximal ppm value for 13C NMR}
}
\description{
Set the PPM bounds for proton (1H) and carbon (13C) to consider in the processing step and then to store in the specMat object
}
