% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{sliceSpectrum}
\alias{sliceSpectrum}
\title{sliceSpectrum}
\usage{
sliceSpectrum(
  spec,
  ppmrange = c(0.5, 9.5),
  flatwidth = 0.004,
  snrfactor = 4,
  maxrange = 0.3,
  excludezones = NULL
)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppmrange}{a ppm range as a list in order to apply the deconvolution}

\item{flatwidth}{specifies the minimum width of a zone in which the spectrum intensities are close to zero to consider this one as a cutting zone (default 0.003 ppm)}

\item{snrfactor}{specifies factor applied on the Std Dev. of the noise used as threshold for first derivate intensities (default=4)}

\item{maxrange}{specifies the maximum width of a cutting zone (default 0.3 ppm)}

\item{excludezones}{specifies the exclusion zones as a matrix (Nx2), each row specifying a zone with 2 columns (ppm min and ppm max) (default NULL)}
}
\value{
a list of ppm range
}
\description{
Slice the spectrum in order to define ranges for Local Deconvolution (LSDeconv)
}
