% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rnmr.R
\name{Spec1r.doProc}
\alias{Spec1r.doProc}
\title{\code{Spec1r.doProc}: generates the real part of a 1H-NMR spectrum from its raw FID}
\usage{
Spec1r.doProc(Input, ...)
}
\arguments{
\item{Input}{A string : absolute path of the Bruker/Varian directory}

\item{procparams}{A list : list of processing parameters, see the  default parameter list \code{Spec1rFromFID.params}}
}
\value{
a Spectrum object
}
\description{
\code{Spec1r.doProc}: generates the real part of a 1H-NMR spectrum from its raw FID
}

