% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnmr.R
\name{Spec1rDoProc}
\alias{Spec1rDoProc}
\title{Spec1rDoProc}
\usage{
Spec1rDoProc(Input, param = Spec1rProcpar)
}
\arguments{
\item{Input}{Full directory path of the raw spectrum}

\item{param}{a Spec1rProcpar list}
}
\value{
spec object
}
\description{
\code{Spec1rDoProc} belongs to the low-level functions group - it processes only one raw spectrum at time.
}
