% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnmr.R
\name{Spec1rFinalize}
\alias{Spec1rFinalize}
\title{Spec1rFinalize}
\usage{
Spec1rFinalize(spec, ...)
}
\arguments{
\item{spec}{object}

\item{...}{undocumented}
}
\value{
spec object
}
\description{
\code{Spec1rFinalize}  finalizes the processing.
}
