% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnmr.R
\docType{data}
\name{Spec1rProcpar}
\alias{Spec1rProcpar}
\title{Spec1rProcpar}
\format{An object of class \code{list} of length 25.}
\usage{
Spec1rProcpar
}
\description{
Initialize Parameter Lists by the default ones
}
\keyword{datasets}
