% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnmr.R
\name{Spec1rReadSpecMatrix}
\alias{Spec1rReadSpecMatrix}
\title{Spec1rReadSpecMatrix}
\usage{
Spec1rReadSpecMatrix(filepack)
}
\arguments{
\item{filepack}{the full path of binary file}
}
\value{
specMat specMat object - int: the Matrix of Spectrum : 1 row <=> 1 spectrum, 1 column <=> a same value of ppm
}
\description{
\code{Spec1rReadSpecMatrix}  Read a Matrix of Spectrum in a binary mode (PACK format)
}
