% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnmr.R
\name{Spec1rWriteSpecMatrix}
\alias{Spec1rWriteSpecMatrix}
\title{Spec1rWriteSpecMatrix}
\usage{
Spec1rWriteSpecMatrix(specmat, ...)
}
\arguments{
\item{specmat}{the Matrix of Spectrum : 1 row <=> 1 spectrum, 1 column <=> a same value of ppm}

\item{...}{undocumented}
}
\description{
\code{Spec1rWriteSpecMatrix} belongs to the low-level functions group - it Write a Matrix of Spectrum in a binary mode (PACK format)
}
