% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{specModel}
\alias{specModel}
\title{specModel}
\usage{
specModel(spec, ppmrange, peaks)
}
\arguments{
\item{spec}{a 'spec' object}

\item{ppmrange}{a ppm range as a list in order to apply the deconvolution}

\item{peaks}{a matrix defining peaks, one peak by row, with columns defined as : pos, ppm, amp, sigma, eta, integral}
}
\value{
a vector with the same size of the spectrum
}
\description{
\code{specModel} belongs to the low-level functions group for deconvolution.
}
