% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{cleanPeaks}
\alias{cleanPeaks}
\title{cleanPeaks}
\usage{
cleanPeaks(spec, peaks, ratioPN, keeprows = FALSE)
}
\arguments{
\item{spec}{a 'spec' object}

\item{peaks}{a data.frame of the input peaks}

\item{ratioPN}{Threshold of the Peak/Noise ratio below which the peaks will be rejected}

\item{keeprows}{indicates if row names must be preserved.}
}
\value{
a data.frame of the remaining peaks
}
\description{
\code{cleanPeaks} cleans the peaks under a specified threshold and also remove redundant peaks having the same position
}
