% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvTools.R
\name{filterByWT}
\alias{filterByWT}
\title{filterByWT}
\usage{
filterByWT(s, wavelet, type = 0)
}
\arguments{
\item{s}{the spectral signal as a numerical vector}

\item{wavelet}{the type of the wavelet: 0 for HAAR, 1 for DAUB2, 2 for DAUB4, 3 for DAUB8}

\item{type}{the type of the threshold :  0 for Soft threshold, 1 for Hard threshold}
}
\value{
a vector of the same dimension as the entry one
}
\description{
\code{filterByWT} applies a filtering based on wavelet using the universal threshold
}
