% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{data}
\name{LOGFILE}
\alias{LOGFILE}
\title{LOGFILE}
\format{An object of class \code{character} of length 1.}
\usage{
LOGFILE
}
\description{
Default log file - Default value = stderr()
}
\keyword{datasets}
