% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_func.R
\name{Lorentz}
\alias{Lorentz}
\title{Lorentz}
\usage{
Lorentz(ppm, amp, x0, sigma)
}
\arguments{
\item{ppm}{a vector of ppm values}

\item{amp}{amplitude of the lorentzian}

\item{x0}{central value of the lorentzian}

\item{sigma}{half-width of the lorentzian}
}
\value{
a vector of the lorentzian values (same size as ppm)
}
\description{
\code{Lorentz} belongs to the low-level functions group for deconvolution.
}
